refactor: move client utility functions to client-utils.ts and update imports
